v0.5.2 optimized NLG6 detection and HV power calculation

+++ ED_BMSdiag/_CS_dfs.h
@@ -16,9 +16,9 @@
 //--------------------------------------------------------------------------------
 //! \file    CS_dfs.h
 //! \brief   Definitions and structures for the Cooling- & Subsystems module.
-//! \date    2016-July
-//! \author  My-Lab-odyssey
-//! \version 0.3.0
+//! \date    2016-October
+//! \author  MyLab-odyssey
+//! \version 0.4.0
 //--------------------------------------------------------------------------------
 #ifndef CS_DFS_H
 #define CS_DFS_H
@@ -26,19 +26,19 @@
 //CS data structure
 typedef struct {       
 
-  unsigned int  CoolingTemp;      //!< main cooling temperatur measurement / 8
+  uint16_t  CoolingTemp;      //!< main cooling temperatur measurement / 8
   byte CoolingPumpTemp;           //!< temperature at cooling pump, offset 50
   byte CoolingPumpLV;             //!< 12V onboard voltage of cooling pump / 10
-  unsigned int CoolingPumpAmps;   //!< current measured at cooling pump / 5
+  uint16_t CoolingPumpAmps;   //!< current measured at cooling pump / 5
   byte CoolingPumpRPM;            //!< RPM in % of cooling pump (value / 255 * 100%)
-  unsigned int CoolingPumpOTR;    //!< operating time record of cooling pump in hours
+  uint16_t CoolingPumpOTR;    //!< operating time record of cooling pump in hours
   byte CoolingFanRPM;             //!< RPM in % of cooling fan (value / 255 * 100%)
-  unsigned int CoolingFanOTR;     //!< operating time record of cooling fan in hours
-  unsigned int BatteryHeaterOTR;  //!< operating time record of PTC heater of battery
+  uint16_t CoolingFanOTR;     //!< operating time record of cooling fan in hours
+  uint16_t BatteryHeaterOTR;  //!< operating time record of PTC heater of battery
   byte BatteryHeaterON;           //!< Status of the PTC heater of the battery
   unsigned long VaccumPumpOTR;    //!< operating time record of vaccum pump
-  int VaccumPumpPress1;           //!< pressure of vaccum pump measuerement #1 in mbar?
-  int VaccumPumpPress2;           //!< pressure of vaccum pump measuerement #2 in mbar?
+  int16_t VaccumPumpPress1;           //!< pressure of vaccum pump measuerement #1 in mbar?
+  int16_t VaccumPumpPress2;           //!< pressure of vaccum pump measuerement #2 in mbar?
 
 } CoolingSub_t;
